Add rainbowModeOn to colorBtn, eraserBtn, remove it, console.log and eraserBtn.style color setting from watchColorChange
- Simplified button functions for better UX because previous setup was complicated and included the eraser button taking
the current color and changing its background color to imply to the user that it can be used to set the current color back on click.
- The reason for using this option was due to the lack of space in the UI, but it got resolved with redesign of the UI (adding of the
color button and putting the clear button in the same line as the grid slider.

- Benefits from doing this are: controls that are used more are now placed better, while the clear button is placed out of other
controls/buttons row, reducing the accidental clicking while drawing.